requirements.txt
discord-py-interactions==4.3.4
requests==2.22.0
ics==0.7
discord==1.7.3
python-dotenv==0.18.0
python_dateutil==2.8.1
table2ascii
